pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.7.1",
    "flask>=3.1.3",
    "matplotlib>=3.10.8",
    "pymongo[srv]",
    "pillow",
    "aiohttp",
    "psycopg2-binary",
]
